requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==36.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.11.0
django-heroku==0.3.1
django-mail-queue==3.2.5
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.0.4
idna==3.3
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
boto3==1.26.24
sentry-sdk==1.12.1
